demo-version/requirements.txt
# Core dependencies
instantneo[openai]>=0.1.1
python-dotenv>=1.0.0

# Streamlit and UI
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Database connectivity
pyodbc>=4.0.39

# Development and testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0